backend/requirements/base.txt
# 族谱系统基础依赖包
# 该文件包含了所有环境共用的Python包依赖

# Django核心框架
Django==4.2.7
django-ninja==1.0.1

# Dajngo组件
django_extensions==4.1.0

# 数据库相关
mysqlclient==2.2.7
django-redis==5.4.0

# 配置管理
python-decouple==3.8

# 日志处理
loguru==0.7.2

# 异步任务队列
celery==5.3.4
redis==5.0.1

# 图像处理
Pillow==10.1.0

# 时间处理
python-dateutil==2.8.2

# HTTP请求
requests==2.31.0

# JSON Web Token
PyJWT==2.8.0

# 密码加密
bcrypt==4.1.2

# CORS处理
django-cors-headers==4.3.1

# 数据验证
pydantic==2.5.0

# 文件类型检测
python-magic==0.4.27

# 字符串处理
python-slugify==8.0.1

# 环境变量
python-dotenv==1.0.0

# 二维码生成
qrcode==7.4.2